Rename organisation components
